move linter count converter
