--- src/DynamoCoreWpf/Controls/Converters/LinterIssueCountToVisibilityConverter.cs
@@ -1,33 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Data;
-
-namespace Dynamo.UI.Controls.Converters
-{
-    /// <summary>
-    /// Converts an integer (linter issues count) to a visibility state
-    /// </summary>
-    [ValueConversion(typeof(int), typeof(Visibility))]
-    internal class LinterIssueCountToVisibilityConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (!(value is int issueCount) || issueCount == 0)
-            {
-                return Visibility.Collapsed;
-            }
-
-            return Visibility.Visible;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
