+++ src/DynamoCoreWpf/UI/Converters.cs
@@ -3161,4 +3161,25 @@ namespace Dynamo.Controls
         }
     }
 
+    /// <summary>
+    /// Converts an integer (linter issues count) to a visibility state
+    /// </summary>
+    [ValueConversion(typeof(int), typeof(Visibility))]
+    internal class LinterIssueCountToVisibilityConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (!(value is int issueCount) || issueCount == 0)
+            {
+                return Visibility.Collapsed;
+            }
+
+            return Visibility.Visible;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
 }
